requirements.txt
streamlit
pandas
databricks-sql-connector
python-docx
pypandoc